add appointment time check (in working hours)

+++ controllers/appointments.js
@@ -97,6 +97,16 @@ exports.addAppointment = async (req, res, next) => {
       });
     }
 
+    if (req.body.apptDate) {
+      const reqDate = new Date(req.body.apptDate);
+      if (reqDate < shop.openTime || reqDate > shop.closeTime) {
+        return res.status(400).json({
+          succes: false,
+          message: "the apptDate is not in the massage shop working hours",
+        });
+      }
+    }
+
     const appointment = await Appointment.create(req.body);
 
     return res.status(200).json({
@@ -118,13 +128,6 @@ exports.updateAppointment = async (req, res, next) => {
   try {
     let appointment = await Appointment.findById(req.params.id);
 
-    if (!appointment) {
-      return res.status(404).json({
-        success: false,
-        message: `No appointment with the id of ${req.params.id}`,
-      });
-    }
-
     if (
       appointment.user.toString() !== req.user.id &&
       req.user.role !== "admin"
@@ -133,6 +136,23 @@ exports.updateAppointment = async (req, res, next) => {
         success: false,
         message: `User ${req.user.id} is not authorized to update this appointment`,
       });
+    }
+
+    if (!appointment) {
+      return res.status(404).json({
+        success: false,
+        message: `No appointment with the id of ${req.params.id}`,
+      });
+    }
+
+    if (req.body.apptDate) {
+      const reqDate = new Date(req.body.apptDate);
+      if (reqDate < shop.openTime || reqDate > shop.closeTime) {
+        return res.status(400).json({
+          succes: false,
+          message: "the apptDate is not in the massage shop working hours",
+        });
+      }
     }
 
     appointment = await Appointment.findByIdAndUpdate(req.params.id, req.body, {
@@ -160,13 +180,6 @@ exports.deleteAppointment = async (req, res, next) => {
   try {
     const appointment = await Appointment.findById(req.params.id);
 
-    if (!appointment) {
-      return res.status(404).json({
-        success: false,
-        message: `No appointment with the id of ${req.params.id}`,
-      });
-    }
-
     if (
       appointment.user.toString() !== req.user.id &&
       req.user.role !== "admin"
@@ -176,6 +189,14 @@ exports.deleteAppointment = async (req, res, next) => {
         message: `User ${req.user.id} is not authorized to delete this appointment`,
       });
     }
+
+    if (!appointment) {
+      return res.status(404).json({
+        success: false,
+        message: `No appointment with the id of ${req.params.id}`,
+      });
+    }
+
     await appointment.deleteOne();
 
     return res.status(200).json({
